pyproject.toml
[tool.poetry]
name = "space-telegram"
version = "0.1.0"
description = ""
authors = ["Andrey Miskov <amiskov@gmail.com>"]
readme = "README.md"
packages = [{include = "space_telegram"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
environs = "^9.5.0"
python-telegram-bot = "^13"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
